fixed more warnings (#1748)
* fixed more warnings

* Potential bug, in PETSIRD if x and y axis are the same size then z won't be checked.

* fix hidden base function

* Use of format with typeid to avoid using outdated `warning("string %s", "bla")`

+++ src/scatter_buildblock/sample_scatter_points.cxx
@@ -31,7 +31,7 @@ static inline float
 random_point(const float low, const float high)
 {
   /* returns a pseudo random number which holds in the bounds low and high */
-  const float result = (rand() * (high - low)) / RAND_MAX + low;
+  const float result = static_cast<float>((static_cast<double>(rand()) * (high - low)) / RAND_MAX + low);
   assert(low <= result);
   assert(high >= result);
   return result;
